document_translator/requirements.txt
google-generativeai
typer
rich
python-docx
pdf2docx
tenacity
python-dotenv
EbookLib
beautifulsoup4